+++ EstudantePosGraduacao.hpp
@@ -13,12 +13,11 @@ class EstudantePosGraduacao : public Estudante {
     public:
         EstudantePosGraduacao(const std::string& _nome, const std::string &_cpf,const std::string& _data_de_nascimento, const std::string& _email, const std::string& _senha, const std::string &_matricula, const std::string &_curso, const std::string &_tipoPos, const std::string &_linhaDePesquisa);
         ~EstudantePosGraduacao();
-        int get_valorRU() const;
+        double get_valorRU() const;
         std::string get_tipoPos() const;
         std::string get_linhaDePesquisa() const;
         int get_prazoDeDevolucao() const override;
         
-        //(thales) == o metodo set_valorRU é dessa classe ou de outra acima? -- vou deixar já implementado... tirem se não precisar
         void set_tipoPos(std::string _tipoPos);
         void set_linhaDePesquisa(std::string _linhaDePesquisa);
         static void set_valorRU(double _valorRU);
